PHP 8 Deprecated: Implicit conversion from float (#217)
This fixes #217

+++ include/configclass.php
@@ -60,7 +60,7 @@ function mergesort(&$array, $cmp_function = 'strcmp') {
 	if (count($array) < 2) return;
 
 	// Split the array in half
-	$halfway = count($array) / 2;
+	$halfway = ((int) (count($array) / 2));
 	$array1 = array_slice($array, 0, $halfway);
 	$array2 = array_slice($array, $halfway);
